requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
opencv-python>=4.10.0.82
pydantic>=2.7.3
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
ultralytics
#labelImg>=1.8.6
ruamel.yaml==0.18.6